# benchmarks/locustfile.py
#!/usr/bin/env python3
"""
==============================================================================
SALESTORM ARFA: 10,000-Contender High-Throughput Locust Benchmark
SysCrafters Hackathon - Performance & Chaos Benchmark Suite
==============================================================================
Architecture:
  - Models 10,000 concurrent contenders storming the limited flash drop SKU.
  - Custom tags: "buy_now", "payment", "telemetry", "chaos"
  - Injects 2% duplicate clicks to validate distributed idempotency barrier.
  - Tracks custom P99 latency, failure rates, throughput (RPS), and zero-oversell invariant.
==============================================================================
"""

import json
import logging
import os
import random
import time
import uuid
from typing import Dict, List, Any

from locust import HttpUser, task, between, tag, events
from locust.runners import MasterRunner, WorkerRunner

# ============================================================================
# Benchmark Metrics Registry
# ============================================================================
class ContentionMetrics:
    def __init__(self):
        self.granted_leases = 0       # HTTP 201 Created
        self.stock_exhausted = 0      # HTTP 409 Conflict
        self.duplicate_hits = 0       # Idempotent deduplication cache hits
        self.payments_confirmed = 0   # HTTP 200 Order confirmed
        self.payments_declined = 0    # HTTP 402 Card declined
        self.unexpected_errors = 0
        self.latencies_ms: List[float] = []
        self.start_time = time.time()

    def record_reservation(self, status_code: int, latency_ms: float, is_duplicate: bool):
        self.latencies_ms.append(latency_ms)
        if status_code == 201:
            if is_duplicate:
                self.duplicate_hits += 1
            else:
                self.granted_leases += 1
        elif status_code == 409:
            self.stock_exhausted += 1
        else:
            self.unexpected_errors += 1

    def calculate_percentiles(self) -> Dict[str, float]:
        if not self.latencies_ms:
            return {"p50": 0.0, "p95": 0.0, "p99": 0.0, "max": 0.0}
        sorted_lats = sorted(self.latencies_ms)
        n = len(sorted_lats)
        return {
            "p50": round(sorted_lats[int(n * 0.50)], 2),
            "p95": round(sorted_lats[min(int(n * 0.95), n - 1)], 2),
            "p99": round(sorted_lats[min(int(n * 0.99), n - 1)], 2),
            "max": round(sorted_lats[-1], 2)
        }

metrics = ContentionMetrics()

# ============================================================================
# Locust Contender User Class
# ============================================================================
class FlashSaleContender(HttpUser):
    """
    Simulates a high-intent shopper attempting to claim a flash sale unit.
    Fires at sub-millisecond intervals during the flash drop window.
    """
    # Ultra-low wait time modeling high-contention flash storm
    wait_time = between(0.002, 0.02)

    def on_start(self):
        """Initializes user session and distinct UUIDv4 Idempotency Key."""
        self.user_id = str(uuid.uuid4())
        self.session_idempotency_key = f"usr_{uuid.uuid4()}"
        self.active_reservation_id = None
        self.has_purchased = False

    @tag("buy_now")
    @task(10)
    def attempt_reservation(self):
        """
        Attempts atomic reservation lease against /api/v1/checkout/reserve.
        Injects a 2% duplicate click rate to stress-test the distributed idempotency barrier.
        """
        if self.has_purchased or self.active_reservation_id:
            return

        is_duplicate = random.random() < 0.02
        # Use existing key for duplicates, or generate fresh UUIDv4
        idempotency_key = self.session_idempotency_key if is_duplicate else f"req_{uuid.uuid4()}"

        headers = {
            "Content-Type": "application/json",
            "Idempotency-Key": idempotency_key
        }
        payload = {
            "product_id": "flash_sku_titanium_01",
            "quantity": 1
        }

        t0 = time.perf_counter()
        with self.client.post(
            "/api/v1/checkout/reserve",
            json=payload,
            headers=headers,
            name="/api/v1/checkout/reserve",
            catch_response=True
        ) as response:
            latency_ms = (time.perf_counter() - t0) * 1000.0

            if response.status_code == 201:
                # Successful reservation lease acquired
                metrics.record_reservation(201, latency_ms, is_duplicate)
                data = response.json()
                self.active_reservation_id = data.get("reservation_id")
                response.success()
            elif response.status_code == 409:
                # Clean HTTP 409 drop: Inventory exhausted (Expected behaviour under contention)
                metrics.record_reservation(409, latency_ms, is_duplicate)
                response.success()
            else:
                metrics.record_reservation(response.status_code, latency_ms, is_duplicate)
                response.failure(f"Unexpected status code: {response.status_code}")

    @tag("payment")
    @task(5)
    def complete_payment(self):
        """
        Simulates payment capture for acquired leases against /api/v1/payments/execute.
        95% of payments succeed; 5% decline simulating card networks with compensating release.
        """
        if not self.active_reservation_id or self.has_purchased:
            return

        succeeds = random.random() < 0.95
        headers = {
            "Content-Type": "application/json",
            "Idempotency-Key": f"pay_{self.active_reservation_id}"
        }
        payload = {
            "reservation_id": self.active_reservation_id,
            "payment_method_token": "tok_visa_4242" if succeeds else "tok_visa_declined",
            "success": succeeds
        }

        with self.client.post(
            "/api/v1/payments/execute",
            json=payload,
            headers=headers,
            name="/api/v1/payments/execute",
            catch_response=True
        ) as response:
            if response.status_code == 200:
                metrics.payments_confirmed += 1
                self.has_purchased = True
                self.active_reservation_id = None
                response.success()
            elif response.status_code in (402, 404, 410):
                # 402: Card declined (reclaimed by Lua release)
                metrics.payments_declined += 1
                self.active_reservation_id = None
                response.success()
            else:
                response.failure(f"Payment execution failed: {response.status_code}")

    @tag("telemetry")
    @task(1)
    def query_telemetry_hud(self):
        """Queries live telemetry counters mirroring frontend mission control stream."""
        self.client.get("/api/v1/telemetry", name="/api/v1/telemetry")


# ============================================================================
# Event Hooks: Custom SRE Benchmark Report
# ============================================================================
@events.test_start.add_listener
def on_test_start(environment, **kwargs):
    logging.info("=" * 80)
    logging.info("SALESTORM ARFA: 10,000 CONTENDER FLASH SALE BENCHMARK INITIALIZED")
    logging.info("Target Host:      %s", environment.host)
    logging.info("Target Contenders: 10,000 Contenders")
    logging.info("Catalog SKU:      flash_sku_titanium_01 (Cap: 100 units)")
    logging.info("=" * 80)


@events.test_stop.add_listener
def on_test_stop(environment, **kwargs):
    duration = max(time.time() - metrics.start_time, 0.001)
    total_reqs = len(metrics.latencies_ms)
    throughput = round(total_reqs / duration, 1)
    pcts = metrics.calculate_percentiles()

    print("\n" + "=" * 85)
    print(" SALESTORM ARFA: SRE PERFORMANCE & INVARIANT BENCHMARK REPORT")
    print("=" * 85)
    print(f" Total Ingress Demanded:       {total_reqs:,} Requests")
    print(f" Wall-Clock Duration:          {duration:.2f} Seconds")
    print(f" Sustained Ingress Throughput: {throughput:,} RPS")
    print(f" Granted Leases (HTTP 201):    {metrics.granted_leases} (Catalog Cap: 100)")
    print(f" Clean Drops (HTTP 409):       {metrics.stock_exhausted:,}")
    print(f" Duplicate Requests Prevented: {metrics.duplicate_hits} (Idempotency Barrier Verified)")
    print(f" Payments Confirmed:           {metrics.payments_confirmed}")
    print(f" Payments Declined (Released): {metrics.payments_declined}")
    print(f" Unexpected Errors:            {metrics.unexpected_errors}")
    print("-" * 85)
    print(f" P50 Latency:                  {pcts['p50']} ms")
    print(f" P95 Latency:                  {pcts['p95']} ms")
    print(f" P99 Latency:                  {pcts['p99']} ms  [Target SLO: < 50.0 ms]")
    print(f" Max Latency:                  {pcts['max']} ms")
    print("=" * 85)

    if metrics.granted_leases <= 100:
        print(" >>> [AUDIT RESULT: PASS] ZERO OVERSELLING INVARIANT PRESERVED. EXACT 100 UNITS CEILING.")
    else:
        print(f" >>> [AUDIT RESULT: FAIL] INVARIANT BREACH! {metrics.granted_leases} units granted (Over by {metrics.granted_leases - 100}).")
    print("=" * 85 + "\n")